moved SP selectors to their own file
