+++ aimmd/distributed/pathmovers.py
@@ -36,6 +36,7 @@ from asyncmd.trajectory.propagate import (
 from asyncmd.utils import ensure_mdconfig_options
 
 from ._config import _SEMAPHORES
+from .spselectors import SPSelector, RCModelSPSelector
 
 
 logger = logging.getLogger(__name__)
@@ -383,7 +384,7 @@ class TwoWayShootingPathMover(ModelDependentPathMover):
         # cut and concatenate returns (traj_to_state, first_state_reached)
         # but since we already know about the states we ignore them here
         (fw_traj, _), (bw_traj, _) = concats
-        # NOTE: this is actually not necessary as we alreay load our own
+        # NOTE: this is actually not necessary as we already load our own
         # private copy of the model at the beginning of this move
         # load the selecting model for accept/reject
         #model = self.get_model(stepnum=stepnum)
@@ -458,141 +459,3 @@ class TwoWayShootingPathMover(ModelDependentPathMover):
                           accepted=accepted,
                           p_acc=p_acc,
                           )
-
-
-#TODO: DOCUMENT
-class RCModelSPSelector:
-    def __init__(self, scale=1., distribution="lorentzian",
-                 density_adaptation=True, exclude_first_last_frame=True):
-        self.distribution = distribution
-        self.scale = scale
-        self.density_adaptation = density_adaptation
-        # whether we allow to choose first and last frame
-        # if False they will also not contribute to sum_bias and accept/reject
-        self.exclude_first_last_frame = exclude_first_last_frame
-
-    @property
-    def distribution(self):
-        """Return the name of the shooting point selection distribution."""
-        return self._distribution
-
-    @distribution.setter
-    def distribution(self, val):
-        if val.lower() == 'gaussian':
-            self._f_sel = self._gaussian
-            self._distribution = val
-        elif val.lower() == 'lorentzian':
-            self._f_sel = self._lorentzian
-            self._distribution = val
-        else:
-            raise ValueError('Distribution must be one of: '
-                             + '"gaussian" or "lorentzian"')
-
-    def _lorentzian(self, z):
-        return self.scale / (self.scale**2 + z**2)
-
-    def _gaussian(self, z):
-        return np.exp(-z**2/self.scale)
-
-    async def f(self, snapshot, trajectory, model):
-        """Return the unnormalized proposal probability of a snapshot."""
-        # we expect that 'snapshot' is a len 1 trajectory!
-        z_sel = await model.z_sel(snapshot)
-        any_nan = np.any(np.isnan(z_sel))
-        if any_nan:
-            logger.error('The model predicts NaNs. '
-                         + 'We used np.nan_to_num to proceed')
-            z_sel = np.nan_to_num(z_sel)
-        # casting to float makes errors when the np-array is not size-1,
-        # i.e. we check that snapshot really was a len-1 trajectory
-        ret = float(self._f_sel(z_sel))
-        if self.density_adaptation:
-            committor_probs = await model(snapshot)
-            if any_nan:
-                committor_probs = np.nan_to_num(committor_probs)
-            density_fact = model.density_collector.get_correction(
-                                                            committor_probs
-                                                                  )
-            ret *= float(density_fact)
-        if ret == 0.:
-            if await self.sum_bias(trajectory) == 0.:
-                logger.error("All SP weights are 0. Using equal probabilities.")
-                return 1.
-        return ret
-
-    async def probability(self, snapshot, trajectory, model):
-        """Return proposal probability of the snapshot for this trajectory."""
-        # we expect that 'snapshot' is a len 1 trajectory!
-        sum_bias = await self.sum_bias(trajectory, model)
-        if sum_bias == 0.:
-            logger.error("All SP weights are 0. Using equal probabilities.")
-            if self.exclude_first_last_frame:
-                return 1. / (len(trajectory) - 2)
-            else:
-                return 1. / len(trajectory)
-        return (await self.f(snapshot, trajectory, model)) / sum_bias
-
-    async def sum_bias(self, trajectory, model):
-        """
-        Return the partition function of proposal probabilities for trajectory.
-        """
-        biases = await self._biases(trajectory, model)
-        return np.sum(biases)
-
-    async def _biases(self, trajectory, model):
-        z_sels = await model.z_sel(trajectory)
-        any_nan = np.any(np.isnan(z_sels))
-        if any_nan:
-            logger.error('The model predicts NaNs. '
-                         + 'We used np.nan_to_num to proceed')
-            z_sels = np.nan_to_num(z_sels)
-        ret = self._f_sel(z_sels)
-        if self.density_adaptation:
-            committor_probs = await model(trajectory)
-            if any_nan:
-                committor_probs = np.nan_to_num(committor_probs)
-            density_fact = model.density_collector.get_correction(
-                                                            committor_probs
-                                                                  )
-            ret *= density_fact.reshape(committor_probs.shape)
-        if self.exclude_first_last_frame:
-            ret = ret[1:-1]
-        return ret
-
-    async def pick(self, trajectory, model):
-        """Return the index of the chosen snapshot within trajectory."""
-        # NOTE: this does not register the SP with model!
-        #       i.e. we do stuff different than in the ops selector
-        #       For the distributed case we need to save the predicted
-        #       commitment probabilities at the shooting point with the MCStep
-        #       this way we can make sure that they are added to the central model
-        #       in the same order as the shooting results to the trainset
-        biases = await self._biases(trajectory, model)
-        sum_bias = np.sum(biases)
-        if sum_bias == 0.:
-            logger.error('Model not able to give educated guess.\
-                         Choosing based on luck.')
-            # we can not give any meaningfull advice and choose at random
-            if self.exclude_first_last_frame:
-                # choose from [1, len(traj) - 1 )
-                return np.random.randint(1, len(trajectory) - 1)
-            else:
-                # choose from [0, len(traj) )
-                return np.random.randint(len(trajectory))
-
-        # if self.exclude_first_last_frame == True
-        # biases will be have the length of traj - 2,
-        # i.e. biases already excludes the two frames
-        # that means the idx we choose here is shifted by one in that case,
-        # e.g. idx=0 means frame_idx=1 in the trajectory
-        # (and we can not choose the last frame because biases ends before)
-        rand = np.random.random() * sum_bias
-        idx = 0
-        prob = biases[0]
-        while prob <= rand and idx < len(biases) - 1:
-            idx += 1
-            prob += biases[idx]
-        # and return chosen idx
-        if self.exclude_first_last_frame:
-            idx += 1
-        return idx
